refactor(db): rename keys database definition file
