+++ apps/db/project.json
@@ -26,7 +26,7 @@
       "executor": "nx:run-commands",
       "dependsOn": ["install", "keys-up"],
       "options": {
-        "command": "poetry run python {projectRoot}/src/db/pg_keys.py"
+        "command": "poetry run python {projectRoot}/src/db/keys.py"
       }
     },
     "init-mongo": {
